requirements.txt
# AI Raitha Mitra - Python Dependencies

# Core Framework
flask==3.1.2
flask-cors==6.0.1

# Machine Learning
tensorflow==2.20.0
pillow==11.3.0
numpy==2.3.3

# AI/ML APIs
google-generativeai==0.8.5

# Security
werkzeug==3.1.3

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Database (SQLite is built into Python)
# No additional packages needed for SQLite

# Note: All advanced features (Chat, Farm Planner, Yield Prediction, 
# Financial Health, Community Network, Map) use the above dependencies.
# No additional packages are required.